+++ core/commons/enclave_manager/enclave_manager.go
@@ -7,18 +7,23 @@ package enclave_manager
 
 import (
 	"context"
+	"fmt"
 	"github.com/docker/docker/client"
-	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_bindings"
-	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
 	"github.com/kurtosis-tech/kurtosis/commons/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/commons/docker_network_allocator"
+	"github.com/kurtosis-tech/kurtosis/commons/enclave_manager/enclave_context"
 	"github.com/kurtosis-tech/kurtosis/commons/object_name_providers"
 	"github.com/kurtosis-tech/kurtosis/initializer/api_container_launcher"
-	"github.com/labstack/gommon/log"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
-	"google.golang.org/grpc"
 	"net"
+	"time"
+)
+
+const (
+	// The API container is responsible for disconnecting/stopping everything in its network when stopped, so we need
+	//  to give it some time to do so
+	apiContainerStopTimeout = 10 * time.Second
 )
 
 // Manages Kurtosis enclaves, and creates new ones in response to running tasks
@@ -36,10 +41,15 @@ type EnclaveManager struct {
 
 // NOTE: thisContainerId is the ID of the container in which this code is executing, so that it can be mounted inside
 //  the new enclave such that it can communicate with the API container
-func (manager *EnclaveManager) CreateEnclave(ctx context.Context, thisContainerId string, enclaveId string, log *logrus.Logger, isPartitioningEnabled bool) (*EnclaveContext, error) {
+func (manager *EnclaveManager) CreateEnclave(
+		setupCtx context.Context,
+		log *logrus.Logger,
+		thisContainerId string,
+		enclaveId string,
+		isPartitioningEnabled bool) (*enclave_context.EnclaveContext, error) {
 	dockerManager := docker_manager.NewDockerManager(log, manager.dockerClient)
 
-	matchingNetworks, err := dockerManager.GetNetworkIdsByName(ctx, enclaveId)
+	matchingNetworks, err := dockerManager.GetNetworkIdsByName(setupCtx, enclaveId)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred finding enclaves with name '%v', which is necessary to ensure that our enclave doesn't exist yet", enclaveId)
 	}
@@ -49,9 +59,11 @@ func (manager *EnclaveManager) CreateEnclave(ctx context.Context, thisContainerI
 
 	enclaveObjNameProvider := object_name_providers.NewEnclaveObjectNameProvider(enclaveId)
 
+	teardownCtx := context.Background()  // Separate context for tearing stuff down in case the input context is cancelled
+
 	log.Debugf("Creating Docker network for enclave '%v'...", enclaveId)
 	networkId, networkIpAndMask, gatewayIp, freeIpAddrTracker, err := manager.dockerNetworkAllocator.CreateNewNetwork(
-		ctx,
+		setupCtx,
 		dockerManager,
 		log,
 		enclaveId,
@@ -63,25 +75,55 @@ func (manager *EnclaveManager) CreateEnclave(ctx context.Context, thisContainerI
 		//  networks with our network name and delete them
 		return nil, stacktrace.Propagate(err, "An error occurred allocating a new network for enclave '%v'", enclaveId)
 	}
+	shouldDeleteNetwork := true
+	defer func() {
+		if shouldDeleteNetwork {
+			if err := dockerManager.RemoveNetwork(teardownCtx, networkId); err != nil {
+				log.Errorf("Creating the enclave didn't complete successfully, so we tried to delete network '%v' that we created but an error was thrown:")
+				fmt.Fprintln(log.Out, err)
+				log.Errorf("ACTION REQUIRED: You'll need to manually remove network with ID '%v'!!!!!!!", networkId)
+			}
+		}
+	}()
 	log.Debugf("Docker network '%v' created successfully with ID '%v' and subnet CIDR '%v'", enclaveId, networkId, networkIpAndMask.String())
 
-	// TODO Mount other containers (i.e. initializer) inside the network
+	log.Debugf("Connecting this container to the enclave network so that it can interact with the containers in the enclave...")
+	mountIp, err := freeIpAddrTracker.GetFreeIpAddr()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting a free IP for mounting this container inside the enclave")
+	}
+	if err := dockerManager.ConnectContainerToNetwork(setupCtx, networkId, thisContainerId, mountIp); err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred connecting this container to the enclave network")
+	}
+	shouldDisconnectThisContainer := true
+	defer func() {
+		if shouldDisconnectThisContainer {
+			if err := dockerManager.DisconnectContainerFromNetwork(teardownCtx, thisContainerId, networkId); err != nil {
+				log.Errorf("Creating the enclave didn't complete successfully, so we tried to disconnect this container from enclave network but an error was thrown:")
+				fmt.Fprintln(log.Out, err)
+				log.Errorf("ACTION REQUIRED: You'll need to manually disconnect container with ID '%v' from network with ID '%v'!!!!!!!", thisContainerId, networkId)
+			}
+		}
+	}()
+	log.Debugf("Successfully connected this container to the enclave network so that it can interact with the containers in the enclave")
 
 	// TODO use hostnames rather than IPs, which makes things nicer and which we'll need for Docker swarm support
 	// We need to create the IP addresses for BOTH containers because the testsuite needs to know the IP of the API
 	//  container which will only be started after the testsuite container
-	kurtosisApiIp, err := freeIpAddrTracker.GetFreeIpAddr()
+	apiContainerIpAddr, err := freeIpAddrTracker.GetFreeIpAddr()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting an IP for the Kurtosis API container")
 	}
 
-	if err := dockerManager.CreateVolume(ctx, enclaveId); err != nil {
+	if err := dockerManager.CreateVolume(setupCtx, enclaveId); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating enclave volume '%v'", enclaveId)
 	}
+	// NOTE: We could defer a deletion of this volume unless the function completes successfully - right now, Kurtosis
+	//  doesn't do any volume deletion
 
 	apiContainerName := enclaveObjNameProvider.ForApiContainer()
 	apiContainerId, err := manager.apiContainerLauncher.Launch(
-		ctx,
+		setupCtx,
 		log,
 		dockerManager,
 		apiContainerName,
@@ -89,72 +131,41 @@ func (manager *EnclaveManager) CreateEnclave(ctx context.Context, thisContainerI
 		networkId,
 		networkIpAndMask.String(),
 		gatewayIp,
-		kurtosisApiIp,
+		apiContainerIpAddr,
 		[]net.IP{},  // TODO Add the other containers that we mount in here
 		isPartitioningEnabled,
+		map[string]bool{thisContainerId: true},
 	)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred launching the API container")
 	}
-
-	grpc.Dial()
-	apiClient := kurtosis_core_rpc_api_bindings.NewApiContainerServiceClient()
-	networkCtx := networks.NewNetworkContext()
-
-
-
-}
-
-/*
-Helper function for making a best-effort attempt at removing a network and the containers inside after the task is
-	done running against the enclave
-	exited (either normally or with error)
-*/
-func removeNetworkDeferredFunc(
-	testTeardownContext context.Context,
-	log *logrus.Logger,
-	dockerManager *docker_manager.DockerManager,
-	networkId string,
-	initializerContainerId string) {
-	log.Debugf("Attempting to remove Docker network with ID %v...", networkId)
-	containerIds, err := dockerManager.GetContainerIdsConnectedToNetwork(testTeardownContext, networkId)
-	if err != nil {
-		errorDesc := fmt.Sprintf("An error occurred getting the containers connected to network '%v' so we can stop them:", networkId)
-		logErrorAndRecommendManualIntervention(log, errorDesc, err, networkId)
-		return
-	}
-
-	for _, containerId := range containerIds {
-		if containerId == initializerContainerId {
-			// We don't want to kill the initializer since it could be coordinating other tests, but we need it gone
-			//  from the network before we can delete the network
-			if err := dockerManager.DisconnectContainerFromNetwork(testTeardownContext, initializerContainerId, networkId); err != nil {
-				errorDesc := fmt.Sprintf("An error occurred disconnecting the initializer container from the network, which prevents the network from being deleted:")
-				logErrorAndRecommendManualIntervention(log, errorDesc, err, networkId)
-				return
-			}
-		} else {
-			if err := dockerManager.KillContainer(testTeardownContext, containerId); err != nil {
-				errorDesc := fmt.Sprintf("An error occurred killing container '%v', which prevents the network from being deleted:", containerId)
-				logErrorAndRecommendManualIntervention(log, errorDesc, err, networkId)
-				return
+	// The API container is started successfully and it will disconnect/stop everything in its network when it shuts down,
+	//  so it takes over the responsibility of disconnecting this container
+	shouldDisconnectThisContainer = false
+	shouldStopApiContainer := true
+	defer func() {
+		if shouldStopApiContainer {
+			if err := dockerManager.StopContainer(teardownCtx, apiContainerId, apiContainerStopTimeout); err != nil {
+				log.Errorf("Creating the enclave didn't complete successfully, so we tried to stop the API container but an error was thrown:")
+				fmt.Fprintln(log.Out, err)
+				log.Errorf("ACTION REQUIRED: You'll need to manually stop API container with ID '%v'", apiContainerId)
 			}
 		}
-	}
+	}()
 
-	// Give a tiny bit of time for the container-kills to complete before removing the network
-	time.Sleep(networkTeardownWaitTime)
+	result := enclave_context.NewEnclaveContext(
+		networkId,
+		enclaveId,
+		networkIpAndMask,
+		apiContainerIpAddr,
+	)
 
-	if err := dockerManager.RemoveNetwork(testTeardownContext, networkId); err != nil {
-		errorDesc := fmt.Sprintf("An error occurred removing Docker network with ID %v:", networkId)
-		logErrorAndRecommendManualIntervention(log, errorDesc, err, networkId)
-		return
-	}
-	log.Debugf("Successfully removed Docker network with ID %v", networkId)
+	// Everything started successfully, so the responsibility of deleting the network is now transferred to the caller
+	shouldDeleteNetwork = false
+	shouldStopApiContainer = false
+	return result, nil
 }
 
-func logErrorAndRecommendManualIntervention(log *logrus.Logger, humanReadableErrorDesc string, err error, networkId string) {
-	log.Errorf(humanReadableErrorDesc)
-	log.Error(err.Error())
-	log.Errorf("ACTION REQUIRED: You'll need to delete any remaining containers and the network with ID '%v' manually!!!", networkId)
+func DestroyEnclave(ctx context.Context, enclaveId string) {
+
 }